+++ openeo/rest/vectorcube.py
@@ -7,6 +7,7 @@ from openeo.internal.graph_building import PGNode
 from openeo.internal.warnings import legacy_alias
 from openeo.metadata import CollectionMetadata
 from openeo.rest._datacube import _ProcessGraphAbstraction
+from openeo.rest.mlmodel import MlModel
 from openeo.rest.job import BatchJob
 from openeo.util import dict_no_none
 
@@ -70,7 +71,12 @@ class VectorCube(_ProcessGraphAbstraction):
             }
         )
 
+    def execute(self) -> dict:
+        """Executes the process graph of the imagery."""
+        return self._connection.execute(self.flat_graph())
+
     def download(self, outputfile: str, format: str = "GeoJSON", options: dict = None):
+        # TODO: only add save_result, when not already present (see DataCube.download)
         cube = self.save_result(format=format, options=options)
         return self._connection.download(cube.flat_graph(), outputfile)
 
@@ -114,3 +120,88 @@ class VectorCube(_ProcessGraphAbstraction):
         return self._connection.create_job(process_graph=shp.flat_graph(), additional=job_options)
 
     send_job = legacy_alias(create_job, name="send_job", since="0.10.0")
+
+    @openeo_process
+    def fit_class_random_forest(
+        self,
+        # TODO #279 #293: target type should be `VectorCube` (with adapters for GeoJSON FeatureCollection, GeoPandas, ...)
+        target: dict,
+        # TODO #293 max_variables officially has no default
+        max_variables: Optional[int] = None,
+        num_trees: int = 100,
+        seed: Optional[int] = None,
+    ) -> MlModel:
+        """
+        Executes the fit of a random forest classification based on the user input of target and predictors.
+        The Random Forest classification model is based on the approach by Breiman (2001).
+
+        .. warning:: EXPERIMENTAL: not generally supported, API subject to change.
+
+        :param target: The training sites for the classification model as a vector data cube. This is associated with the target
+            variable for the Random Forest model. The geometry has to be associated with a value to predict (e.g. fractional
+            forest canopy cover).
+        :param max_variables: Specifies how many split variables will be used at a node. Default value is `null`, which corresponds to the
+            number of predictors divided by 3.
+        :param num_trees: The number of trees build within the Random Forest classification.
+        :param seed: A randomization seed to use for the random sampling in training.
+
+        .. versionadded:: 0.16.0
+            Originally added in version 0.10.0 as :py:class:`DataCube <openeo.rest.datacube.DataCube>` method,
+            but moved to :py:class:`VectorCube` in version 0.16.0.
+        """
+        pgnode = PGNode(
+            process_id="fit_class_random_forest",
+            arguments=dict_no_none(
+                predictors=self,
+                # TODO #279 strictly per-spec, target should be a `vector-cube`, but due to lack of proper support we are limited to inline GeoJSON for now
+                target=target,
+                max_variables=max_variables,
+                num_trees=num_trees,
+                seed=seed,
+            ),
+        )
+        model = MlModel(graph=pgnode, connection=self._connection)
+        return model
+
+    @openeo_process
+    def fit_regr_random_forest(
+        self,
+        # TODO #279 #293: target type should be `VectorCube` (with adapters for GeoJSON FeatureCollection, GeoPandas, ...)
+        target: dict,
+        # TODO #293 max_variables officially has no default
+        max_variables: Optional[int] = None,
+        num_trees: int = 100,
+        seed: Optional[int] = None,
+    ) -> MlModel:
+        """
+        Executes the fit of a random forest regression based on training data.
+        The Random Forest regression model is based on the approach by Breiman (2001).
+
+        .. warning:: EXPERIMENTAL: not generally supported, API subject to change.
+
+        :param target: The training sites for the regression model as a vector data cube.
+            This is associated with the target variable for the Random Forest model.
+            The geometry has to associated with a value to predict (e.g. fractional forest canopy cover).
+        :param max_variables: Specifies how many split variables will be used at a node. Default value is `null`, which corresponds to the
+            number of predictors divided by 3.
+        :param num_trees: The number of trees build within the Random Forest classification.
+        :param seed: A randomization seed to use for the random sampling in training.
+
+        .. versionadded:: 0.16.0
+            Originally added in version 0.10.0 as :py:class:`DataCube <openeo.rest.datacube.DataCube>` method,
+            but moved to :py:class:`VectorCube` in version 0.16.0.
+        """
+        # TODO #279 #293: `fit_class_random_forest` should be defined on VectorCube instead of DataCube
+        pgnode = PGNode(
+            process_id="fit_regr_random_forest",
+            arguments=dict_no_none(
+                predictors=self,
+                # TODO #279 strictly per-spec, target should be a `vector-cube`, but due to lack of proper support we are limited to inline GeoJSON for now
+                target=target,
+                max_variables=max_variables,
+                num_trees=num_trees,
+                seed=seed,
+            ),
+        )
+        model = MlModel(graph=pgnode, connection=self._connection)
+        return model